requirements.txt
streamlit
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
mysql-connector-python
SQLAlchemy
pytube
numexpr
unstructured
youtube_transcript_api
validators==0.28.1
python-multipart
arxiv
wikipedia
ddgs
fastapi
uvicorn
pydantic